pyproject.toml
[tool.poetry]
name = "softdesksupport"
version = "0.1.0"
description = ""
authors = ["Myryos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
beautifulsoup4 = "4.12.3"
click = "8.1.7"
colorama = "0.4.6"
cssbeautifier = "1.14.11"
django = "4.2.9"
django-bootstrap-v5 = "1.0.11"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
djlint = "1.34.1"
editorconfig = "0.12.3"
flake8 = "7.0.0"
html-tag-names = "0.1.2"
html-void-elements = "0.1.0"
jsbeautifier = "1.14.11"
json5 = "0.9.14"
mccabe = "0.7.0"
pathspec = "0.12.1"
pillow = "10.1.0"
pycodestyle = "2.11.1"
pyflakes = "3.2.0"
pyjwt = "2.8.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
regex = "2023.12.25"
six = "1.16.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
tomli = "2.0.1"
tqdm = "4.66.1"
typing-extensions = "4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
